Extract cred_def and schema from /handlers to /libindy
