--- aries_vcx/src/handlers/issuance/schema/mod.rs
@@ -1,1 +1,0 @@
-pub mod schema;
