+++ aries_vcx/src/libindy/mod.rs
@@ -1,2 +1,4 @@
 pub mod proofs;
 pub mod utils;
+pub mod credential_def;
+pub mod schema;
