+++ aries_vcx/src/libindy/schema/mod.rs
@@ -1,5 +1,5 @@
 use crate::error::prelude::*;
-use crate::handlers::issuance::credential_def::PublicEntityStateType;
+use crate::libindy::credential_def::PublicEntityStateType;
 use crate::libindy::utils::anoncreds;
 use crate::utils::constants::DEFAULT_SERIALIZE_VERSION;
 use crate::utils::serialization::ObjectWithVersion;
